requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-appconf==1.0.3
django-ckeditor==5.7.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-model-utils==3.2.0
django-mptt==0.10.0
django-notifications-hq==1.5.0
django-taggit==1.1.0
django-widget-tweaks==1.4.5
idna==2.8
jsonfield==2.0.2
oauthlib==3.1.0
pilkit==2.0
Pillow==6.1.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
